refactor: move common code to helper file
move network generation exception to helpers
