+++ lib/syskit.rb
@@ -127,6 +127,7 @@ require "syskit/connection_graph"
 require "syskit/actual_data_flow_graph"
 require "syskit/data_flow"
 require "syskit/connection_graphs"
+require "syskit/network_generation_exception_helpers"
 require "syskit/exceptions"
 require "syskit/network_generation"
 require "syskit/runtime"
